Minor fixes, updated Unit Tests.
The Header object of the RawNotification Payload no longer has the 'appName' field since it can not actually be set by the NotificationCompat.Builder object. In line with that change the headerText field -- the short String to the right of the application name -- is now displayed correctly when set.

Finally, additional unit tests have been added for the Header and Actions objects of the RawNotification Payloads.