Additional Notification Anatomy documentation
The Notification object is fairly complex and the Android documentation does not do an adequate job of describing the elements within a notification and referencing the appropriate builder functions. The additions to the README form the initial contribution to documenting the notification layouts with their corresponding Notify/NotificationCompat.Builder functions.