Unit tests and bug fixes.
Unit tests for the Metadata/Stackable scoped function have been added. Additionally, a bug that was discovered as a result has been patched; it was discovered that a single stackable notification would take on the click hander for the scenario in which the notification was stacked (i.e multiple notifications with the same stack key).
Various members of the NotifyExtender object have been made visible for testing and the helper functions to access the notification bundle now take in a notification extras bundle as opposed to the notification builder so that the extended notification extras can be accessed.