Notification Icon and color tests updated.
Previously, the notification header tests were using the 'describeContents()' member of the Icon object to compare notification icons -- this was discovered to be the incorrect method since the describeContents function always returns 0, and as a result the assertion would always evalutate to true even if the images were seperate. Additionally, there was a intermittent error with the assertion of the color of the notification, it is gone now but will be investigated further if it does appear again.