Alerting Scoped function, Notif. Channel refactoring.
A new scoped function has been added to support notification alerting configurations such as sounds, vibrations, led colors, and more. While those changes are blocked by an Notification Channel compatibility design problem the configuration is planned to live within the Alerts object.
Additionally, registerChannel has been refactored into a seperate class to keeps the functionality seperate from the notification interop logic which is specific to the notification object.

Other changes include updates to the README title styling as well as enabling the ability to skip testing and linting on the sample project which currently has no such functionality. There should be a noticable increase in build speeds as a result.