Extend alerting to include Sounds, LED, vibration.
The alerting object is now closely related to the NotificationChannels. For newer versions of Android the alerting object is used to construct the NotificationChannel, while on older versions it is used in the Notification builder to assign to the appropriate values. It defines a coupling with channels that will provide the best means of building notification alerts.

Developer feedback will provide additional insights into this particular section of the API